+++ packages/oc-template-esm-compiler/scaffold/src/src/App.tsx
@@ -1,24 +1,25 @@
-import { useState } from 'react';
-import { serverClient, InitialData, ActionOutput } from 'oc-server';
+import { serverClient, type ActionOutput, getInitialData } from 'oc-server';
+import type { Remix } from '@remix-run/dom';
+import { press } from '@remix-run/events/press';
 import styles from './styles.module.css';
 import logo from '../public/logo.png';
 
 type AdditionalData = ActionOutput<'funFact'>;
 
-const App: React.FC<InitialData> = (props: InitialData) => {
-  const { firstName, lastName, born, hobbies } = props;
-  const [additionalData, setAdditionalData] = useState<AdditionalData | null>(
-    null
-  );
-  const [error, setError] = useState('');
+export default function App(this: Remix.Handle) {
+  const { firstName, lastName, born, hobbies } = getInitialData();
+  let additionalData: AdditionalData | null = null;
+  let error = '';
 
   const getFunFact = async () => {
-    setError('');
+    error = '';
     try {
       const data = await serverClient.funFact({ year: born });
-      setAdditionalData(data);
+      additionalData = data;
     } catch (err) {
-      setError(String(err));
+      error = String(err);
+    } finally {
+      this.update();
     }
   };
 
@@ -26,7 +27,7 @@ const App: React.FC<InitialData> = (props: InitialData) => {
     return <div>Something wrong happened!</div>;
   }
 
-  return (
+  return () => (
     <div className={styles.container}>
       <img width="50" height="50" src={logo} alt="Logo" />
       <h1 style={{ margin: '0 0 20px 0' }}>
@@ -40,11 +41,9 @@ const App: React.FC<InitialData> = (props: InitialData) => {
         </div>
       </div>
       {additionalData && <div>{additionalData.funFact}</div>}
-      <button className={styles.button} onClick={getFunFact}>
+      <button className={styles.button} on={press(getFunFact)}>
         Fun year fact
       </button>
     </div>
   );
-};
-
-export default App;
+}