+++ packages/oc-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # oc-server
 
+## 3.2.1
+
+### Patch Changes
+
+- Add signal option to serverClient and getAction
+
 ## 3.2.0
 
 ### Minor Changes