+++ packages/oc-template-elm-compiler/CHANGELOG.md
@@ -1,5 +1,13 @@
 # oc-template-elm-compiler
 
+## 2.1.3
+
+### Patch Changes
+
+- Add signal option to serverClient and getAction
+- Updated dependencies
+  - oc-vite-compiler@5.1.3
+
 ## 2.1.2
 
 ### Patch Changes