+++ packages/oc-template-react-compiler/CHANGELOG.md
@@ -1,5 +1,13 @@
 # oc-template-react-compiler
 
+## 10.1.3
+
+### Patch Changes
+
+- Add signal option to serverClient and getAction
+- Updated dependencies
+  - oc-vite-compiler@5.1.3
+
 ## 10.1.2
 
 ### Patch Changes